update search css, add date clear button, restructure
update search css, add date clear button

added initial search wildcard to caputlog accordian to search for pvName* to include searching for pvs with sufixes

restructure caputlog page to be a directory under components
