+++ src/components/caputlog/CaputLogSearchLayout.css
@@ -1,0 +1,46 @@
+/* Override search box submit button to match MUI contained button style */
+.sui-search-box__submit {
+  background: #00334d !important;
+  color: #fff;
+  border: none;
+  border-radius: 4px;
+  box-shadow: 0px 2px 2px rgba(25, 118, 210, 0.08);
+  font-weight: 500;
+  font-family: Ubuntu,-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,"Helvetica Neue",Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol";
+  font-size: 0.875rem;
+  text-transform: uppercase;
+  padding: 6px 16px;
+  min-width: 64px;
+  transition: background 0.2s;
+}
+.sui-search-box__submit:hover, .sui-search-box__submit:focus {
+  background: #002335;
+}
+
+.sui-search-box__wrapper {
+  position: relative;
+  display: inline-block;
+  box-sizing: border-box;
+  width: 100%;
+  font-family: Ubuntu,-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,"Helvetica Neue",Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol";
+  font-size: 1rem;
+}
+
+.sui-search-box__text-input {
+  width: 100%;
+  padding: 8px 12px;
+  border: 1px solid #0000003b; /* Match MUI outlined input border */
+  border-radius: 4px;
+  min-height: inherit;
+  box-shadow: none;
+  font-size: inherit;
+  font-family: inherit;
+  font-weight: 400;
+}
+
+.sui-search-box__wrapper:hover {
+  /* Match react-datepicker input hover border color */
+  box-shadow: 0 0 0 2px #000000de;
+  min-height: inherit;
+  border-radius: 4px;
+}
