+++ src/components/caputlog/index.js
@@ -1,0 +1,2 @@
+import CaputLogPage from "./CaputLogPage";
+export default CaputLogPage;
